+++ src/main/java/com/ohgiraffers/adregamdi/cityanddong/query/application/service/CityQueryService.java
@@ -17,7 +17,6 @@ public class CityQueryService {
     public CityDTO findCityByCityName(String cityName) {
 
         City result = cityQueryRepository.findCityByCityName(cityName);
-
         return new CityDTO(
                 result.getCityNo(),
                 result.getCityName()
