+++ src/main/java/com/ohgiraffers/adregamdi/data/command/application/service/DataService.java
@@ -105,10 +105,12 @@ public class DataService {
     public CategoryDTO insertCategory(String categoryName) {
 
         CategoryDTO findResult = dataAPIService.findCategoryByCategoryName(categoryName);
+        System.out.println("findResult = " + findResult);
         if (findResult != null) {
             return findResult;
         }
         CategoryDTO insertedCategory = dataAPIService.insertCategory(categoryName);
+        System.out.println("insertedCategory = " + insertedCategory);
         return insertedCategory;
     }
 
@@ -126,6 +128,7 @@ public class DataService {
     public DongDTO insertDong(int dongCode, String dongName) {
 
         DongDTO findResult = dataAPIService.findDongByDongName(dongName);
+        System.out.println("findResult = " + findResult);
         if (findResult != null) {
             return findResult;
         }
