+++ src/main/java/com/ohgiraffers/adregamdi/place/query/application/service/PlaceQueryService.java
@@ -1,8 +1,8 @@
 package com.ohgiraffers.adregamdi.place.query.application.service;
 
+import com.ohgiraffers.adregamdi.category.query.application.service.CategoryQueryService;
 import com.ohgiraffers.adregamdi.place.command.application.dto.PlaceDTO;
 import com.ohgiraffers.adregamdi.place.query.domain.entity.SearchPlace;
-import com.ohgiraffers.adregamdi.place.query.domain.service.CategoryQueryService;
 import com.ohgiraffers.adregamdi.place.query.infra.repository.PlaceMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
