+++ src/main/java/com/ohgiraffers/adregamdi/place/query/infra/repository/PlaceQueryRepository.java
@@ -4,6 +4,4 @@ import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceQueryRepository extends JpaRepository<Place, Long> {
-
-    Place findPlaceByPlaceNameAndRoadPlaceAddress(String placeName, String roadPlaceAddress);
 }
