--- src/main/java/com/ohgiraffers/adregamdi/place/query/infra/repository/PlaceRepository.java
@@ -1,28 +1,0 @@
-package com.ohgiraffers.adregamdi.place.query.infra.repository;
-
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Place;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-
-//public interface PlaceRepository{
-
-
-
-
-
-//   Place findPlaceByPlaceNo(Long placeNo);
-    // 맵 id 받아서 -- 서비스로 올리기
-//    public Placee findPlaceReviewList(){}
-    // 카테고리 코드 받아서 조회
-//    public Placee findPlaceListByCategory(){}
-    // 검색 ?? 잘 모르겠음
-//    public Placee findPlaceListBySearch(){}
-    // 검색
-//    public Placee findPlaceListBySearchAndCategory(){}
-// 좌표 값으로 리스트 조회인데 잘 모르겠음
-//    public Placee findPlaceListByBound(){}
-
-//}
