+++ src/test/java/com/ohgiraffers/adregamdi/place/query/application/service/PlaceQueryServiceTests.java
@@ -28,5 +28,4 @@ public class PlaceQueryServiceTests {
 
     }
 
-
 }
